Move geme rendering to GameRenderComponent
By moving geme rendering out from GameScreen and into a new GuiComponent sub-class, we can leave GameScreen as the default Screen, while also adding other screens using the game rendering in other ways.
