+++ src/de/gurkenlabs/litiengine/gui/screens/Screen.java
@@ -10,26 +10,8 @@ import de.gurkenlabs.litiengine.gui.GuiComponent;
  * 
  */
 public abstract class Screen extends GuiComponent {
-  // The screen's layer determines the render order.
-  // The higher the value, the later the screen will be rendered.
-  // This is useful if you wish to make some sort of overlay, such as a UI.
-  private int layer;
-
   protected Screen(final String screenName) {
-    this(screenName, 0);
-  }
-
-  protected Screen(final String screenName, int layer) {
     super(0,0);
     this.setName(screenName);
-    this.setLayer(layer);
-  }
-
-  public int getLayer() {
-    return layer;
-  }
-
-  public void setLayer(int layer) {
-    this.layer = layer;
   }
 }
